@@ -34,13 +34,12 @@
     "country-emoji": "^1.5.6",
     "country-flag-icons": "^1.5.5",
     "fuse.js": "^6.6.2",
-    "moment-timezone": "^0.5.34",
-    "worldcities": "^0.1.7"
+    "moment-timezone": "^0.5.34"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^7.32.0",
+    "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.3.0",
     "prettier": "^2.5.1",
     "typescript": "^4.4.3"
